# ExpenseMate/ExpenseMate/src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/room": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "getRoomDetails",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "controller"
        ],
        "operationId": "addUsersToRoom",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/room/settle": {
      "post": {
        "tags": [
          "controller"
        ],
        "operationId": "settleExpenses",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/room/create": {
      "post": {
        "tags": [
          "controller"
        ],
        "operationId": "createRoom",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      }
    },
    "/expense": {
      "get": {
        "tags": [
          "controller"
        ],
        "operationId": "getExpense",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "controller"
        ],
        "operationId": "addExpense",
        "parameters": [
          {
            "name": "roomId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string"
          },
          "roomName": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "Expense": {
        "type": "object",
        "properties": {
          "expenseId": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "paidBy": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "shares": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Share"
            }
          }
        }
      },
      "Share": {
        "type": "object",
        "properties": {
          "shareId": {
            "type": "string"
          },
          "expenseId": {
            "type": "string"
          },
          "owedBy": {
            "type": "string"
          },
          "owedTo": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}